reorganization, doc
